Update tests
Move Emit tests into separate file
